// src/modules/users/interfaces/user.interface.ts
export interface User {
  snils: string;
  email: string;
  phoneNumber: string;
  password: string;
  firstName: string;
  lastName: string;
  patronymic: string;
  gender: string;
  dateOfBirth: string;
  city: string;
  street: string;
  houseNumber: number;
  apartmentNumber: number;
  postalCode: number;
  passportSeries: number;
  passportNumber: number;
  whoIssuedPassport: string;
  dateIssueOfThePassport: string;
}